core-1555 Refactor code.
